+++ sample/src/main/kotlin/org/jetbrains/desktop/sample/linux/SkikoCustomTitlebarLinux.kt
@@ -47,7 +47,7 @@ internal class SkikoCustomTitlebarLinux(
     private var titleTextLineCreator = TextLineCreator(cachedFontSize = 0f, cachedText = "")
 
     companion object {
-        const val CUSTOM_TITLEBAR_HEIGHT = 55U
+        const val CUSTOM_TITLEBAR_HEIGHT = 55
         const val BUTTON_LINE_WIDTH: LogicalPixels = 5.0
         const val MOVE_RADIUS: LogicalPixels = 3.0
         val COLOR_DARK_GRAY = Color.makeRGB(128, 128, 128)
@@ -74,23 +74,23 @@ internal class SkikoCustomTitlebarLinux(
         buttonLayout = layout
         rectangles.clear()
         buttonLayout.let {
-            val buttonsLeftWidth = it.layoutLeft.size.toUInt() * BUTTON_SIZE.width
-            val buttonsRightWidth = it.layoutRight.size.toUInt() * BUTTON_SIZE.width
+            val buttonsLeftWidth = it.layoutLeft.size * BUTTON_SIZE.width
+            val buttonsRightWidth = it.layoutRight.size * BUTTON_SIZE.width
             val rect = LogicalRect(
                 x = buttonsLeftWidth,
-                y = 0U,
+                y = 0,
                 width = size.width - buttonsRightWidth - buttonsLeftWidth,
                 height = CUSTOM_TITLEBAR_HEIGHT,
             )
             rectangles.add(Pair(rect, WindowButtonType.Title))
             for ((i, button) in it.layoutLeft.withIndex()) {
-                val rect = LogicalRect(x = i.toUInt() * BUTTON_SIZE.height, y = 0U, width = BUTTON_SIZE.width, height = BUTTON_SIZE.height)
+                val rect = LogicalRect(x = i * BUTTON_SIZE.height, y = 0, width = BUTTON_SIZE.width, height = BUTTON_SIZE.height)
                 rectangles.add(Pair(rect, button))
             }
             for ((i, button) in it.layoutRight.withIndex()) {
                 val rect = LogicalRect(
-                    x = size.width - ((it.layoutRight.size.toUInt() - i.toUInt()) * BUTTON_SIZE.width),
-                    y = 0U,
+                    x = size.width - ((it.layoutRight.size - i) * BUTTON_SIZE.width),
+                    y = 0,
                     width = BUTTON_SIZE.width,
                     height = BUTTON_SIZE.height,
                 )
@@ -176,7 +176,7 @@ internal class SkikoCustomTitlebarLinux(
     }
 
     fun onMouseDown(event: Event.MouseDown): EventHandlerResult {
-        val headerRect = LogicalRect(x = 0U, y = 0U, width = size.width, height = size.height)
+        val headerRect = LogicalRect(x = 0, y = 0, width = size.width, height = size.height)
         return if (headerRect.contains(event.locationInWindow) && event.button == MouseButton.LEFT) {
             leftClickStartLocation = event.locationInWindow
             isDragging = false
@@ -192,7 +192,7 @@ internal class SkikoCustomTitlebarLinux(
         window: Window,
         windowState: WindowState,
     ): EventHandlerResult {
-        val headerRect = LogicalRect(x = 0U, y = 0U, width = size.width, height = size.height)
+        val headerRect = LogicalRect(x = 0, y = 0, width = size.width, height = size.height)
         val leftClickStartWindowButton = leftClickStartLocation?.let { leftClickStartLocation ->
             rectangles.firstOrNull { it.first.contains(leftClickStartLocation) }?.second
         }
@@ -235,7 +235,7 @@ internal class SkikoCustomTitlebarLinux(
     }
 
     fun onMouseMoved(locationInWindow: LogicalPoint, window: Window): EventHandlerResult {
-        val headerRect = LogicalRect(x = 0U, y = 0U, width = size.width, height = size.height)
+        val headerRect = LogicalRect(x = 0, y = 0, width = size.width, height = size.height)
         lastMouseLocation = locationInWindow
         return if (headerRect.contains(locationInWindow) &&
             !isDragging &&
